Inicio Coordinador GUI correcciones
