+++ src/Vistas/InicioAdministradorController.java
@@ -1,27 +1,42 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package Vistas;
 
 import java.net.URL;
 import java.util.ResourceBundle;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Button;
 
-/**
- * FXML Controller class
- *
- * @author israz
- */
 public class InicioAdministradorController implements Initializable {
 
-    /**
-     * Initializes the controller class.
-     */
+    @FXML
+    private Button btnVerAcademias;
+    @FXML
+    private Button btnVerPlanesDeTrabajo;
+    @FXML
+    private Button btnVerMinutas;
+    @FXML
+    private Button btnVerProgramasDeEstudios;
+
     @Override
     public void initialize(URL url, ResourceBundle rb) {
-        // TODO
+        
     }    
+
+    @FXML
+    private void clicVerAcademias(ActionEvent event) {
+    }
+
+    @FXML
+    private void clicVerAcademicos(ActionEvent event) {
+    }
+
+    @FXML
+    private void clicVerExperienciasEducativas(ActionEvent event) {
+    }
+
+    @FXML
+    private void clicVerPeriodos(ActionEvent event) {
+    }
     
 }
