+++ src/Vistas/InicioCoordinadorController.java
@@ -19,6 +19,7 @@ import javafx.stage.Stage;
  */
 public class InicioCoordinadorController implements Initializable {
 
+    @FXML
     private Button btnVerAvancesProgramaticos;
     @FXML
     private Button btnVerPlanesDeTrabajo;
@@ -26,8 +27,6 @@ public class InicioCoordinadorController implements Initializable {
     private Button btnVerMinutas;
     @FXML
     private Button btnVerProgramasDeEstudios;
-    @FXML
-    private Button btnVerAcademias;
 
     /**
      * Initializes the controller class.
@@ -37,18 +36,22 @@ public class InicioCoordinadorController implements Initializable {
         
     }    
 
+    @FXML
     private void clicVerAvancesProgramáticos(ActionEvent event) {
         abrirOpciónCliqueada("VerAvances.fxml");
     }
 
+    @FXML
     private void clicVerPlanesDeTrabajo(ActionEvent event) {
         abrirOpciónCliqueada("VerPlanesDeTrabajo.fxml");
     }
 
+    @FXML
     private void clicVerMinutasDeReunionAcademica(ActionEvent event) {
         abrirOpciónCliqueada("VerMinutas.fxml");
     }
 
+    @FXML
     private void clicVerProgramasDeEstudios(ActionEvent event) {
         abrirOpciónCliqueada("VerPlanesDeCurso.fxml");
     }
@@ -64,19 +67,4 @@ public class InicioCoordinadorController implements Initializable {
         }
     }
 
-    @FXML
-    private void clicVerAcademias(ActionEvent event) {
-    }
-
-    @FXML
-    private void clicVerAcademicos(ActionEvent event) {
-    }
-
-    @FXML
-    private void clicVerExperienciasEducativas(ActionEvent event) {
-    }
-
-    @FXML
-    private void clicVerPeriodos(ActionEvent event) {
-    }
 }
